+++ src/Exception/InvalidRouteActionException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\RadBundle\Exception;
+
+class InvalidRouteActionException extends RadException
+{
+}